Create README.md
On my actual life I`m engineer. Chemistry, but I`m working in aluminium industry. Together with a friend we developed some applications to follow moving products, different stages where the products are. He was in charge for software part, I was in charge for the architecture  of this sistem. It involves more than 30 production facilities and around 600 people.  It starts with one of my many ideas, more then 10 years ago, and soon it becomes reality and it grows very fast. My friend left the company and I intend to develop this applications collection. The project was built on Visual Basic and SQL Server, but I will try to use Java, instead of Visual Basic, because one of the newest application the we started in the last period is based on Java